Separated out SessionBackend and AuthnEvent definitions from sdb.py to avoid circular dependencies.
