+++ src/oic/oauth2/__init__.py
@@ -49,6 +49,7 @@ from oic.utils.http_util import BadRequest
 from oic.utils.http_util import Response
 from oic.utils.http_util import SeeOther
 from oic.utils.keyio import KeyJar
+from oic.utils.session_backend import session_update
 from oic.utils.time_util import utc_time_sans_frac
 
 __author__ = "rohe0002"
@@ -195,6 +196,7 @@ class Client(PBase):
             timeout=timeout,
         )
 
+        self.sso_db = None
         self.client_id = client_id
         self.client_authn_method = client_authn_method
 
@@ -649,8 +651,9 @@ class Client(PBase):
             except KeyError:
                 self.grant[_state] = self.grant_class(resp=resp)
 
-            if 'id_token' in resp:
-                self.sdb.update(_state, "sub", resp["id_token"]["sub"])
+            if "id_token" in resp:
+                session_update(self.sso_db, _state, "sub", resp["id_token"]["sub"])
+                session_update(self.sso_db, _state, "issuer", resp["id_token"]["iss"])
 
         return resp
